<rdar://problem/5813180> 10A35: Stalled NetInstall of 10A35
git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@23582 e92bca22-270c-0410-9cea-e3f1106b6a1c

+++ launchd/src/libvproc_internal.h
@@ -94,7 +94,7 @@ struct logmsg_s {
 		const char *msg;
 		uint64_t msg_offset;
 	};
-	size_t obj_sz;
+	uint64_t obj_sz;
 	char data[0];
 };
